requirements.txt
fastapi
uvicorn
boto3
python-dotenv
SQLAlchemy
email-validator
python-multipart
